requirements.txt
face-recognition
opencv-python
numpy
scikit-learn
pickle-mixin
Pillow
customtkinter
tkinterweb
kagglehub